Sadge Update
I've made progress, but I still have to fx it. Basically I created a parquet file that holds a pre generated graph significantly increasing speed. One problem. The parquet file is taking way too long to create, and the code making it isn't perfect due to edge cases. Need to make it be able to continue where it left off